Add better GUI
