+++ src/main/java/Fredricksen/commands/ByeCommand.java
@@ -1,0 +1,10 @@
+package Fredricksen.commands;
+
+public class ByeCommand extends Command {
+    public ByeCommand() {};
+
+    @Override
+    public String execute() {
+        return "Bye!";
+    }
+}
