+++ src/main/java/Fredricksen/commands/DeadlineCommand.java
@@ -1,0 +1,20 @@
+package Fredricksen.commands;
+
+import Fredricksen.commands.Command;
+import Fredricksen.tasks.TaskList;
+import Fredricksen.tasks.taskType.Deadline;
+public class DeadlineCommand extends Command {
+    public TaskList tasks;
+    private String fullCommand;
+    public DeadlineCommand(String fullCommand, TaskList tasks) {
+        this.fullCommand = fullCommand;
+        this.tasks = tasks;
+    }
+
+    @Override
+    public String execute() {
+        Deadline newDlTask = new Deadline(this.fullCommand, "D", false);
+        this.tasks.addTask(newDlTask);
+        return displayTask(newDlTask, tasks);
+    }
+}
