--- src/main/java/Launcher.java
@@ -1,11 +1,0 @@
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
