--- src/main/java/commands/ByeCommand.java
@@ -1,10 +1,0 @@
-package commands;
-
-public class ByeCommand extends Command {
-    public ByeCommand() {};
-
-    @Override
-    public String execute() {
-        return "Bye! See you never! Stop making an old man do the work for you!";
-    }
-}
