--- src/main/java/commands/UnknownCommand.java
@@ -1,12 +1,0 @@
-package commands;
-
-public class UnknownCommand extends Command {
-
-    public UnknownCommand() {}
-
-    @Override
-    public String execute() {
-        return "Sorry Old Man Fredricksen don't recognise this input! "
-                + "Type \"help\" if you need a guide on input format!";
-    }
-}
