+++ src/test/java/ui/UiTest.java
@@ -1,5 +1,6 @@
 package ui;
 
+import Fredricksen.ui.Ui;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
